backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
passlib[argon2]
python-jose[cryptography]
email-validator
python-multipart